chore: Move ConversationIgnoredDeviceSystemMessageCellDescription & ConversationEncryptionInfoSystemMessageCellDescription into their own files
